pyproject.toml
[tool.poetry]
name = "airbnbpriceregression"
version = "0.1.0"
description = ""
authors = ["Nagi0 <brunolefonseca13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
protobuf = "<3.20.*"
requests = "^2.23.0"
python-dotenv = "1.0.1"
kagglehub = "^0.3.6"
polars = "1.20.0"
pyarrow = "^19.0.0"
pandas = "^2.2.3"
numpy = "<2.0"
scikit-learn = "^1.6.1"
seaborn = "^0.13.2"
matplotlib = "^3.10.0"
xgboost = "^2.1.3"
tensorflow = "<2.11"
tensorflow-io-gcs-filesystem = "0.27.0"
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
